+++ lib/assets/javascripts/new-dashboard/components/NotificationCarto3Release.vue
@@ -10,20 +10,23 @@
           class="text"
           v-html="
             $t('Carto3Release.message', {
-              link: `<a href='https://carto.com/blog/announcing-carto-spatial-extension-bigquery/' target='_blank'>${$t('Carto3Release.link')}</a>`,
+              link: `<a href='https://carto.com/signup' target='_blank'>${$t('Carto3Release.link')}</a>`,
+              contact: `<a href='mailto:enterprise-support@carto.com' target='_blank'>${$t('Carto3Release.contactUs')}</a>`
             })
           "
         >
         </span>
       </div>
     </span>
-    <button
+
+    <!-- HACK: Non-closable banner -->
+    <!-- <button
       class="notification__close-button"
       @click="close"
       aria-label="Close"
     >
       <img src="../assets/icons/common/close-black.svg" />
-    </button>
+    </button> -->
   </div>
 </template>